// src/services/transactions/transaction-type.ts
//  enum TransactionStatusEnum {
//   PENDING = 'Pending',
//   PROCESSING = 'Processing',
//   SUCCESS = 'Successful',
//   FAILED = 'Failed',
// }
// ...
export interface TransactionResponse {
  amount: number
  created_at: string
  currency: string
  custom_id: string
  description: string
  environment: string
  failed_reason: FailedReason
  failed_reason_id: string
  fee: Fee
  fee_id: string
  id: string
  merchant: Merchant
  merchant_id: string
  msisdn: string
  provider: string
  reference: string
  status: TransactionStatusEnum
  type: TransactionTypesEnum
  updated_at: string
  used_issuer: UsedIssuer
  used_issuer_id: string
}

interface UsedIssuer {
  created_at: string
  id: string
  issuer_provider_rates: IssuerProviderRate[]
  name: string
  updated_at: string
}

export interface IssuerProviderRate {
  cashin_rate: number
  cashout_rate: number
  created_at: string
  id: string
  issuer: string
  issuer_id: string
  provider: Provider
  provider_id: string
  updated_at: string
}
// ...
 interface FailedReason {
  code: string
  created_at: string
  id: string
  reason: string
  updated_at: string
}

 interface Fee {
  amount: number
  created_at: string
  currency: string
  id: string
  percentage: number
  rate: number
  type: string
  updated_at: string
}

 interface Merchant {
  address: string
  business_name: string
  business_type: string
  email: string
  registration_no: string
  balance: Balance[]
  created_at: string
  id: string
  is_verified: boolean
  key: Key[]
  merchant_charge: MerchantCharge[]
  phone: string
  status: MerchantStatus
  tax_id: string
  updated_at: string
  user: User[]
  webhook: Webhook[]
  website: string
}

 interface Balance {
  amount: number
  balance_histories: BalanceHistory[]
  created_at: string
  currency: string
  id: string
  merchant: Merchant2
  merchant_id: string
  type: string
  updated_at: string
}

 interface BalanceHistory {
  amount: number
  balance: Balance2
  balance_id: string
  created_at: string
  id: string
  reason: string
  type: string
  updated_at: string
}

 interface Balance2 {
  value: string
}

 interface Merchant2 {
  value: string
}

 interface Key {
  created_at: string
  environment: string
  id: string
  key: string
  merchant: Merchant3
  merchant_id: string
  updated_at: string
}

 interface Merchant3 {
  value: string
}

 interface MerchantCharge {
  created_at: string
  id: string
  merchant: Merchant4
  merchantID: string
  payin_charge: number
  payout_charge: number
  provider: Provider
  provider_id: string
  updated_at: string
}

 interface Merchant4 {
  value: string
}

 interface Provider {
  cashin_issuer_used: string[]
  cashout_issuer_used: string[]
  country_iso_code: string
  created_at: string
  currency: string
  default_payin_rate: number
  default_payout_rate: number
  id: string
  image_url: string
  method: string
  operator: string
  updated_at: string
}

 interface User {
  created_at: string
  email: string
  first_name: string
  id: string
  is_active: boolean
  is_verified: boolean
  last_login_at: string
  last_name: string
  merchant: string
  merchant_id: string
  phone: string
  role: string
  updated_at: string
}

 interface Webhook {
  created_at: string
  environment: string
  event: string
  id: string
  merchant: Merchant5
  merchant_id: string
  secret: string
  updated_at: string
  url: string
}